+++ src/main/java/io/lettuce/core/RedisBloomFilterCommandBuilder.java
@@ -8,12 +8,12 @@ package io.lettuce.core;
 
 import java.util.List;
 
-import io.lettuce.core.bf.BfInfoValue;
-import io.lettuce.core.bf.BfInfoValueParser;
+import io.lettuce.core.probabilistic.BfInfoValue;
+import io.lettuce.core.probabilistic.BfInfoValueParser;
 import io.lettuce.core.probabilistic.ScanDumpValue;
 import io.lettuce.core.probabilistic.ScanDumpValueParser;
-import io.lettuce.core.bf.arguments.BfInsertArgs;
-import io.lettuce.core.bf.arguments.BfReserveArgs;
+import io.lettuce.core.probabilistic.arguments.BfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.BfReserveArgs;
 import io.lettuce.core.codec.RedisCodec;
 import io.lettuce.core.output.*;
 import io.lettuce.core.protocol.BaseRedisCommandBuilder;
