+++ src/main/java/io/lettuce/core/api/async/RedisCuckooFilterAsyncCommands.java
@@ -8,10 +8,10 @@ package io.lettuce.core.api.async;
 
 import java.util.List;
 import io.lettuce.core.RedisFuture;
-import io.lettuce.core.cf.CfInfoValue;
+import io.lettuce.core.probabilistic.CfInfoValue;
 import io.lettuce.core.probabilistic.ScanDumpValue;
-import io.lettuce.core.cf.arguments.CfInsertArgs;
-import io.lettuce.core.cf.arguments.CfReserveArgs;
+import io.lettuce.core.probabilistic.arguments.CfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.CfReserveArgs;
 
 /**
  * Asynchronous executed commands for Cuckoo Filter.
