+++ src/main/java/io/lettuce/core/api/reactive/RedisBloomFilterReactiveCommands.java
@@ -7,10 +7,10 @@
 package io.lettuce.core.api.reactive;
 
 import io.lettuce.core.Value;
-import io.lettuce.core.bf.BfInfoValue;
+import io.lettuce.core.probabilistic.BfInfoValue;
 import io.lettuce.core.probabilistic.ScanDumpValue;
-import io.lettuce.core.bf.arguments.BfInsertArgs;
-import io.lettuce.core.bf.arguments.BfReserveArgs;
+import io.lettuce.core.probabilistic.arguments.BfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.BfReserveArgs;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
