+++ src/main/java/io/lettuce/core/api/reactive/RedisCuckooFilterReactiveCommands.java
@@ -6,10 +6,10 @@
  */
 package io.lettuce.core.api.reactive;
 
-import io.lettuce.core.cf.CfInfoValue;
+import io.lettuce.core.probabilistic.CfInfoValue;
 import io.lettuce.core.probabilistic.ScanDumpValue;
-import io.lettuce.core.cf.arguments.CfInsertArgs;
-import io.lettuce.core.cf.arguments.CfReserveArgs;
+import io.lettuce.core.probabilistic.arguments.CfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.CfReserveArgs;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
