+++ src/main/java/io/lettuce/core/api/sync/RedisBloomFilterCommands.java
@@ -7,10 +7,10 @@
 package io.lettuce.core.api.sync;
 
 import java.util.List;
-import io.lettuce.core.bf.BfInfoValue;
+import io.lettuce.core.probabilistic.BfInfoValue;
 import io.lettuce.core.probabilistic.ScanDumpValue;
-import io.lettuce.core.bf.arguments.BfInsertArgs;
-import io.lettuce.core.bf.arguments.BfReserveArgs;
+import io.lettuce.core.probabilistic.arguments.BfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.BfReserveArgs;
 
 /**
  * Synchronous executed commands for Bloom Filter.
