+++ src/main/java/io/lettuce/core/probabilistic/CfInfoValue.java
@@ -4,7 +4,7 @@
  *
  * SPDX-License-Identifier: MIT
  */
-package io.lettuce.core.cf;
+package io.lettuce.core.probabilistic;
 
 import java.util.Map;
