+++ src/main/java/io/lettuce/core/probabilistic/arguments/CfInsertArgs.java
@@ -4,7 +4,7 @@
  *
  * SPDX-License-Identifier: MIT
  */
-package io.lettuce.core.cf.arguments;
+package io.lettuce.core.probabilistic.arguments;
 
 import io.lettuce.core.CompositeArgument;
 import io.lettuce.core.protocol.CommandArgs;
