+++ src/main/kotlin/io/lettuce/core/api/coroutines/RedisBloomFilterCoroutinesCommands.kt
@@ -8,9 +8,9 @@ package io.lettuce.core.api.coroutines
 
 import io.lettuce.core.ExperimentalLettuceCoroutinesApi
 import kotlinx.coroutines.flow.Flow
-import io.lettuce.core.bf.BfInfoValue
-import io.lettuce.core.bf.arguments.BfInsertArgs
-import io.lettuce.core.bf.arguments.BfReserveArgs
+import io.lettuce.core.probabilistic.BfInfoValue
+import io.lettuce.core.probabilistic.arguments.BfInsertArgs
+import io.lettuce.core.probabilistic.arguments.BfReserveArgs
 import io.lettuce.core.probabilistic.ScanDumpValue
 
 /**
