+++ src/main/kotlin/io/lettuce/core/api/coroutines/RedisCuckooFilterCoroutinesCommands.kt
@@ -7,10 +7,10 @@
 package io.lettuce.core.api.coroutines
 
 import io.lettuce.core.ExperimentalLettuceCoroutinesApi
-import io.lettuce.core.cf.CfInfoValue
+import io.lettuce.core.probabilistic.CfInfoValue
 import io.lettuce.core.probabilistic.ScanDumpValue
-import io.lettuce.core.cf.arguments.CfInsertArgs
-import io.lettuce.core.cf.arguments.CfReserveArgs
+import io.lettuce.core.probabilistic.arguments.CfInsertArgs
+import io.lettuce.core.probabilistic.arguments.CfReserveArgs
 
 /**
  * Coroutine executed commands for Cuckoo Filter.
