+++ src/main/templates/io/lettuce/core/api/RedisBloomFilterCommands.java
@@ -8,9 +8,9 @@ package io.lettuce.core.api;
 
 import java.util.List;
 
-import io.lettuce.core.bf.BfInfoValue;
-import io.lettuce.core.bf.arguments.BfInsertArgs;
-import io.lettuce.core.bf.arguments.BfReserveArgs;
+import io.lettuce.core.probabilistic.BfInfoValue;
+import io.lettuce.core.probabilistic.arguments.BfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.BfReserveArgs;
 import io.lettuce.core.probabilistic.ScanDumpValue;
 
 /**
