+++ src/main/templates/io/lettuce/core/api/RedisCuckooFilterCommands.java
@@ -8,10 +8,10 @@ package io.lettuce.core.api;
 
 import java.util.List;
 
-import io.lettuce.core.cf.CfInfoValue;
+import io.lettuce.core.probabilistic.CfInfoValue;
 import io.lettuce.core.probabilistic.ScanDumpValue;
-import io.lettuce.core.cf.arguments.CfInsertArgs;
-import io.lettuce.core.cf.arguments.CfReserveArgs;
+import io.lettuce.core.probabilistic.arguments.CfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.CfReserveArgs;
 
 /**
  * ${intent} for Cuckoo Filter.
