+++ src/test/java/io/lettuce/core/RedisBloomFilterCommandBuilderUnitTests.java
@@ -6,10 +6,10 @@
  */
 package io.lettuce.core;
 
-import io.lettuce.core.bf.BfInfoValue;
+import io.lettuce.core.probabilistic.BfInfoValue;
 import io.lettuce.core.probabilistic.ScanDumpValue;
-import io.lettuce.core.bf.arguments.BfInsertArgs;
-import io.lettuce.core.bf.arguments.BfReserveArgs;
+import io.lettuce.core.probabilistic.arguments.BfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.BfReserveArgs;
 import io.lettuce.core.codec.StringCodec;
 import io.lettuce.core.protocol.Command;
 import io.netty.buffer.ByteBuf;
