+++ src/test/java/io/lettuce/core/bf/RedisBloomFilterReactiveIntegrationTests.java
@@ -11,8 +11,8 @@ import javax.inject.Inject;
 import io.lettuce.core.Value;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.reactive.RedisReactiveCommands;
-import io.lettuce.core.bf.arguments.BfInsertArgs;
-import io.lettuce.core.bf.arguments.BfReserveArgs;
+import io.lettuce.core.probabilistic.arguments.BfInsertArgs;
+import io.lettuce.core.probabilistic.arguments.BfReserveArgs;
 import io.lettuce.test.ReactiveSyncInvocationHandler;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
